Moved factory classes to individual files
